Refactor: 링크페이지 폴더 공유 기능 커스텀 훅 생성 - #52
