+++ src/lib/hooks/links/index.ts
@@ -1,1 +1,2 @@
 export * from './useHorizontalScroll';
+export * from './useFolderShare';
